jms: Move APICondition out of jms component
This class is not related to jms component
